# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "App-Api",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/tasks": {
            "get": {
                "tags": [
                    "Tasks"
                ],
                "summary": "Get all Tasks",
                "operationId": "be68905d866e856db6c738172b8d929a",
                "responses": {
                    "200": {
                        "description": "A list of all Tasks"
                    }
                }
            },
            "post": {
                "tags": [
                    "Tasks"
                ],
                "summary": "Create a new task",
                "operationId": "ebbaabc58e3db61374a57c53544959b2",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "description",
                                    "columnId",
                                    "userId"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "New Task"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Task description goes here"
                                    },
                                    "columnId": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "userId": {
                                        "type": "integer",
                                        "example": 2
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Task created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "name": {
                                            "type": "string",
                                            "example": "New Task"
                                        },
                                        "description": {
                                            "type": "string",
                                            "example": "Task description goes here"
                                        },
                                        "columnId": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "created_at": {
                                            "type": "string",
                                            "example": "2024-01-01T12:00:00"
                                        },
                                        "updated_at": {
                                            "type": "string",
                                            "example": "2024-01-01T12:00:00"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error"
                    }
                }
            }
        },
        "/api/tasks/{id}": {
            "get": {
                "tags": [
                    "Tasks"
                ],
                "summary": "Get task by ID",
                "operationId": "32c40e2d1129b80c1dcabd538df63a33",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the task to retrieve",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Task retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "name": {
                                            "type": "string",
                                            "example": "Task 1"
                                        },
                                        "description": {
                                            "type": "string",
                                            "example": "This is task description."
                                        },
                                        "columnId": {
                                            "type": "integer",
                                            "example": 101
                                        },
                                        "created_at": {
                                            "type": "string",
                                            "example": "2024-01-01T12:00:00"
                                        },
                                        "updated_at": {
                                            "type": "string",
                                            "example": "2024-01-01T12:00:00"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Task not found"
                    }
                }
            },
            "put": {
                "tags": [
                    "Tasks"
                ],
                "summary": "Update task by ID",
                "operationId": "c36c959b634c99e1e02ad2f3f2c36a2c",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the task to update",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "description",
                                    "columnId",
                                    "userId"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Updated Task"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Updated task description"
                                    },
                                    "columnId": {
                                        "type": "integer",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Task updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "name": {
                                            "type": "string",
                                            "example": "Updated Task"
                                        },
                                        "description": {
                                            "type": "string",
                                            "example": "Updated task description"
                                        },
                                        "columnId": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "created_at": {
                                            "type": "string",
                                            "example": "2024-01-01T12:00:00"
                                        },
                                        "updated_at": {
                                            "type": "string",
                                            "example": "2024-01-01T12:00:00"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Task not found"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Tasks"
                ],
                "summary": "Delete task by ID",
                "operationId": "d81cc7975ea62176b4ef8f74ed95168d",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the task to delete",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Task deleted successfully"
                    },
                    "404": {
                        "description": "Task not found"
                    }
                }
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}